Migrate UserProfile to users app
