+++ resorthub/admin.py
@@ -3,6 +3,4 @@ from __future__ import unicode_literals
 
 from django.contrib import admin
 
-from .models import UserProfile
-
-admin.site.register(UserProfile)
+# Register your models here
