+++ resorthub/forms.py
@@ -3,7 +3,7 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm
 
 from django.contrib.auth.models import User
-from .models import UserProfile
+from users.models import UserProfile
 from resorts.models import SkiPass, Resort
 
 import datetime
