+++ templates/skidom_base.html
@@ -6,10 +6,10 @@
 
     {% if user.is_authenticated %}
         <span class="link"><a href="{% url 'logout' %}">Logout</a></span>
-        <span class="link"><a href="{% url 'usersettings:profile' %}">Profile</a></span>
+        <span class="link"><a href="{% url 'users:profile' %}">Profile</a></span>
     
     {% else %}
-        <span class="link"><a href="{% url 'usersettings:signup' %}">User sign-up</a></span>   
+        <span class="link"><a href="{% url 'users:signup' %}">User sign-up</a></span>   
         <span class="link"><a href="{% url 'login' %}">Login</a></span>   
 
     {% endif %}
