+++ users/admin.py
@@ -3,4 +3,6 @@ from __future__ import unicode_literals
 
 from django.contrib import admin
 
-# Register your models here.
+from .models import UserProfile
+
+admin.site.register(UserProfile)
